Split up the register module
