+++ packages/phoria-islands/src/client/main.ts
@@ -1,6 +1,12 @@
+import {
+	type PhoriaIslandComponentCsrService,
+	type PhoriaIslandCsrMountMode,
+	type PhoriaIslandCsrOptions,
+	csrMountMode
+} from "./csr"
 import type { PhoriaIslandDirective } from "./directives"
 import { PhoriaIsland } from "./phoria-island"
 
-export { PhoriaIsland }
+export { csrMountMode, PhoriaIsland }
 
-export type { PhoriaIslandDirective }
+export type { PhoriaIslandComponentCsrService, PhoriaIslandCsrMountMode, PhoriaIslandCsrOptions, PhoriaIslandDirective }
