--- packages/phoria-islands/src/server/server-entry.ts
@@ -1,8 +1,0 @@
-import type { PhoriaIslandSsrResult } from "~/register"
-import type { PhoriaIsland } from "./phoria-island"
-
-interface PhoriaServerEntry {
-	renderPhoriaIsland: (phoriaIsland: PhoriaIsland<unknown>) => Promise<PhoriaIslandSsrResult>
-}
-
-export type { PhoriaServerEntry }
