+++ packages/phoria-vue/src/server/ssr.ts
@@ -1,5 +1,5 @@
-import { type PhoriaIslandComponentSsrService, type PhoriaIslandProps, createIslandImport } from "@phoria/phoria"
-import type { RenderPhoriaIslandComponent } from "@phoria/phoria"
+import { type PhoriaIslandProps, createIslandImport } from "@phoria/phoria"
+import type { PhoriaIslandComponentSsrService, RenderPhoriaIslandComponent } from "@phoria/phoria/server"
 import { type Component, createSSRApp } from "vue"
 import { renderToWebStream, renderToString } from "vue/server-renderer"
 import { framework } from "~/main"
